From trustedLEn for boolean.
