+++ src/array/boolean/from.rs
@@ -1,0 +1,130 @@
+use crate::buffer::{Bitmap, MutableBitmap};
+
+use super::BooleanArray;
+
+impl BooleanArray {
+    pub fn from_slice<P: AsRef<[bool]>>(slice: P) -> Self {
+        unsafe { Self::from_trusted_len_iter(slice.as_ref().iter().map(Some)) }
+    }
+}
+
+impl BooleanArray {
+    /// Creates a [`BooleanArray`] from an iterator of trusted length.
+    /// # Safety
+    /// The iterator must be [`TrustedLen`](https://doc.rust-lang.org/std/iter/trait.TrustedLen.html).
+    /// I.e. that `size_hint().1` correctly reports its length.
+    #[inline]
+    pub unsafe fn from_trusted_len_iter<I, P>(iter: I) -> Self
+    where
+        P: std::borrow::Borrow<bool>,
+        I: IntoIterator<Item = Option<P>>,
+    {
+        let iterator = iter.into_iter();
+
+        let (validity, values) = trusted_len_unzip(iterator);
+
+        Self::from_data(values, validity)
+    }
+
+    /// Creates a [`PrimitiveArray`] from an falible iterator of trusted length.
+    /// # Safety
+    /// The iterator must be [`TrustedLen`](https://doc.rust-lang.org/std/iter/trait.TrustedLen.html).
+    /// I.e. that `size_hint().1` correctly reports its length.
+    #[inline]
+    pub unsafe fn try_from_trusted_len_iter<E, I, P>(iter: I) -> Result<Self, E>
+    where
+        P: std::borrow::Borrow<bool>,
+        I: IntoIterator<Item = Result<Option<P>, E>>,
+    {
+        let iterator = iter.into_iter();
+
+        let (validity, values) = try_trusted_len_unzip(iterator)?;
+
+        Ok(Self::from_data(values, validity))
+    }
+}
+
+/// Creates a Bitmap and an optional [`Bitmap`] from an iterator of `Option<bool>`.
+/// The first buffer corresponds to a bitmap buffer, the second one
+/// corresponds to a values buffer.
+/// # Safety
+/// The caller must ensure that `iterator` is `TrustedLen`.
+#[inline]
+pub(crate) unsafe fn trusted_len_unzip<I, P>(iterator: I) -> (Option<Bitmap>, Bitmap)
+where
+    P: std::borrow::Borrow<bool>,
+    I: Iterator<Item = Option<P>>,
+{
+    let (_, upper) = iterator.size_hint();
+    let len = upper.expect("trusted_len_unzip requires an upper limit");
+
+    let mut null = MutableBitmap::with_capacity(len);
+    let mut values = MutableBitmap::with_capacity(len);
+
+    for item in iterator {
+        let item = if let Some(item) = item {
+            null.push_unchecked(true);
+            *item.borrow()
+        } else {
+            null.push_unchecked(false);
+            false
+        };
+        values.push_unchecked(item);
+    }
+    assert_eq!(
+        values.len(),
+        len,
+        "Trusted iterator length was not accurately reported"
+    );
+    values.set_len(len);
+    null.set_len(len);
+
+    let bitmap = if null.null_count() > 0 {
+        Some(null.into())
+    } else {
+        None
+    };
+    (bitmap, values.into())
+}
+
+/// # Safety
+/// The caller must ensure that `iterator` is `TrustedLen`.
+#[inline]
+pub(crate) unsafe fn try_trusted_len_unzip<E, I, P>(
+    iterator: I,
+) -> Result<(Option<Bitmap>, Bitmap), E>
+where
+    P: std::borrow::Borrow<bool>,
+    I: Iterator<Item = Result<Option<P>, E>>,
+{
+    let (_, upper) = iterator.size_hint();
+    let len = upper.expect("trusted_len_unzip requires an upper limit");
+
+    let mut null = MutableBitmap::with_capacity(len);
+    let mut values = MutableBitmap::with_capacity(len);
+
+    for item in iterator {
+        let item = if let Some(item) = item? {
+            null.push_unchecked(true);
+            *item.borrow()
+        } else {
+            null.push_unchecked(false);
+            false
+        };
+        values.push_unchecked(item);
+    }
+    assert_eq!(
+        values.len(),
+        len,
+        "Trusted iterator length was not accurately reported"
+    );
+    values.set_len(len);
+    null.set_len(len);
+
+    let bitmap = if null.null_count() > 0 {
+        Some(null.into())
+    } else {
+        None
+    };
+    Ok((bitmap, values.into()))
+}
