+++ src/array/boolean/mod.rs
@@ -1,8 +1,4 @@
-use crate::{
-    buffer::{Bitmap, MutableBitmap},
-    datatypes::DataType,
-    ffi::ArrowArray,
-};
+use crate::{buffer::Bitmap, datatypes::DataType, ffi::ArrowArray};
 
 use super::{ffi::ToFFI, Array, FromFFI};
 
@@ -113,73 +109,7 @@ impl<P: AsRef<[Option<bool>]>> From<P> for BooleanArray {
     }
 }
 
-impl BooleanArray {
-    pub fn from_slice<P: AsRef<[bool]>>(slice: P) -> Self {
-        unsafe { Self::from_trusted_len_iter(slice.as_ref().iter().map(Some)) }
-    }
-}
-
-impl BooleanArray {
-    /// Creates a [`BooleanArray`] from an iterator of trusted length.
-    /// # Safety
-    /// The iterator must be [`TrustedLen`](https://doc.rust-lang.org/std/iter/trait.TrustedLen.html).
-    /// I.e. that `size_hint().1` correctly reports its length.
-    #[inline]
-    pub unsafe fn from_trusted_len_iter<I, P>(iter: I) -> Self
-    where
-        P: std::borrow::Borrow<bool>,
-        I: IntoIterator<Item = Option<P>>,
-    {
-        let iterator = iter.into_iter();
-
-        let (validity, values) = trusted_len_unzip(iterator);
-
-        Self::from_data(values, validity)
-    }
-}
-
-/// Creates a Bitmap and a [`Buffer`] from an iterator of `Option`.
-/// The first buffer corresponds to a bitmap buffer, the second one
-/// corresponds to a values buffer.
-/// # Safety
-/// The caller must ensure that `iterator` is `TrustedLen`.
-#[inline]
-pub(crate) unsafe fn trusted_len_unzip<I, P>(iterator: I) -> (Option<Bitmap>, Bitmap)
-where
-    P: std::borrow::Borrow<bool>,
-    I: Iterator<Item = Option<P>>,
-{
-    let (_, upper) = iterator.size_hint();
-    let len = upper.expect("trusted_len_unzip requires an upper limit");
-
-    let mut null = MutableBitmap::with_capacity(len);
-    let mut values = MutableBitmap::with_capacity(len);
-
-    for item in iterator {
-        let item = if let Some(item) = item {
-            null.push_unchecked(true);
-            *item.borrow()
-        } else {
-            null.push_unchecked(false);
-            false
-        };
-        values.push_unchecked(item);
-    }
-    assert_eq!(
-        values.len(),
-        len,
-        "Trusted iterator length was not accurately reported"
-    );
-    values.set_len(len);
-    null.set_len(len);
-
-    let bitmap = if null.null_count() > 0 {
-        Some(null.into())
-    } else {
-        None
-    };
-    (bitmap, values.into())
-}
-
 mod iterator;
 pub use iterator::*;
+
+mod from;
